已经实现对舵机的shell控制但是效果不好，并不是每次发送都能使得舵机转动

+++ mg_delay.hpp
@@ -27,7 +27,7 @@ public:
     }
     /// @brief 改变角度
     /// @param digree 0-180 digree
-    void digChange(uint16 digree)
+    void digChange(int digree)
     {
         unsigned long up_time_us = (digree * 200 / 18.0) + 500;
         digitalWrite(this->pin, HIGH);
